go back to using bindgen.py
I would've preferred to generate bindings at compile time, even though
building bindgen is really slow. But unfortunately, it requires
libclang, which is a drag on AppVeyor and (more importantly) people's
machines.

Add a test of our own for the size of the param structs, since those are
part of the spec.